# src/features/game_auth/model/response/signin.go
package response

type ResGameSignin struct {
	AccessToken      string `json:"accessToken"`
	RefreshToken     string `json:"refreshToken"`
	UserID           uint   `json:"userID"`
	IsDuplicateLogin bool   `json:"isDuplicateLogin"`
}